relocate getPasswordHashAlgorithm
